backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.1.0
google-generativeai==0.8.5
pydantic[email]==2.11.3
pydantic-settings==2.9.1
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
